Feature: Organized the Key dropdown when adding rules in the admin
Previously, rule classes were in no particular order in the Key dropdown in the admin. This made it very difficult to find a rule when many were available.

This commit adds an optional "category" attribute that can be added to rules. The category is used as an optgroup in the select dropdown in the admin.

Additionally, rules are now in alphabetical order in the admin (within category).