Feature: Cache the Rule model
Rules are configured for a particular model. Hitting the database for
the rules with every transaction can generate a lot of overhead,
especially considering that rules are often configured once, then let
be.

This commit implements Django Cache Machine on the Rule model to reduce
database overhead. This is a breaking change if django-cache-machine is
not available to the project.